fix(reputationdb)!: store categories as integers on disk
The previous commit changed ListMembership.Category to CategoryByte but
left DataType writing it with mmdbtype.String, so every read of a freshly
built database failed with "cannot unmarshal vpn (string) into type
reputationdb.CategoryByte". Nothing that opened a database worked.

Write both category fields as mmdb uint16 instead. mmdb has no 8-bit
integer type, so CategoryByte widens to uint16 to match; the decoder has
a fast path from uint16 into a narrower field either way, and six of the
sixteen bits are in use.

Drop the six is_* booleans and the categories string array from the
record. They said nothing the mask does not, and callers now convert at
their own edges: maat and dump-prefixes build a mask once from their
configured category names and test it with Has, while the gRPC handler
derives the booleans and the names it puts on the wire. The protobuf API
is unchanged.

categoryBits is reordered alphabetically so Strings returns a sorted list
without sorting, which keeps the wire order it had before.

BREAKING CHANGE: the on-disk record schema changed. A reader built from
this commit cannot decode a database built before it, and nothing in the
file marks the difference: database_type is still Techaro-Veil-VPN and
the version a client keys on is a content hash. Cached databases must be
discarded, and database_type wants a bump before this is published.

+++ categories_test.go
@@ -18,8 +18,8 @@ func TestCategoryByteToFromRoundtrip(t *testing.T) {
 			got := FromCategories(st)
 
 			if want != got {
-				t.Logf("want: %x (%v)", byte(want), st)
-				t.Logf(" got: %x (%v)", byte(got), got.Strings())
+				t.Logf("want: %x (%v)", uint16(want), st)
+				t.Logf(" got: %x (%v)", uint16(got), got.Strings())
 				t.Error("did not get expected information in the roundtrip")
 			}
 		})
